refactor(app): extract chat ordering
